TReTA_DAQ/requirements.txt
matplotlib
seaborn
numpy
PyQT5
pyserial